remove Tools submodule
